requirements.txt
flask
pdfplumber
python-docx
pytesseract
Pillow
nltk
scikit-learn